requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudant==2.14.0
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.10
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycryptodome==3.9.8
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
